+++ src/shared/UI/headers/index.js
@@ -1,0 +1,7 @@
+import HeaderActions from './actions';
+import getSelectionHeaderParams from './selection';
+
+export {
+  HeaderActions,
+  getSelectionHeaderParams,
+};